backend/requirements.txt
fastapi==0.115.6
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.0
redis==5.2.1
celery==5.5.1
requests

# numpy==1.24.4
pyttsx3==2.98
groq==0.29.0
pymongo==4.11.3
# numpy==1.24.4